Use jsonrpc and remove bad hearbeating system
The main problem was that nodes had to be checked periodically. Now, Call() and Go() can detect timeout to stop using a node.

The question is to know if JsonRPC is a good idea, or if we should go back to Gob object usage. Anyway, that's not a big work to revert.